Add HikariCP - try with resources

+++ Core/src/main/java/de/tentact/languageapi/spigot/command/LanguageCommand.java
@@ -6,6 +6,9 @@ package de.tentact.languageapi.spigot.command;
 */
 
 import de.tentact.languageapi.AbstractLanguageAPI;
+import de.tentact.languageapi.mysql.MySQL;
+import de.tentact.languageapi.util.I18N;
+import de.tentact.languageapi.util.Source;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandSender;
 import org.bukkit.command.TabExecutor;
@@ -26,12 +29,14 @@ public class LanguageCommand implements TabExecutor {
 
     public static HashMap<Player, List<String>> givenParameter = new HashMap<>();
 
+    private final MySQL mySQL = Source.getMySQL();
+
 
     public boolean onCommand(@NotNull CommandSender commandSender, @NotNull Command command, @NotNull String s, String[] args) {
         if (commandSender instanceof Player) {
             Player player = (Player) commandSender;
             if (player.hasPermission("system.languageapi")) { //lang add lang key MSG | lang remove lang key | lang update lang key msg | lang createlang lang | lang deletelang lang
-                if (args.length >= 2) {
+                if (args.length >= 1) { //lang reload
                     switch (args[0].toLowerCase()) {
                         case "add":
                             if (!(args.length >= 4)) {
@@ -89,30 +94,48 @@ public class LanguageCommand implements TabExecutor {
                                 return false;
                             }
                         case "create":
+                            if(!(args.length >= 2)) {
+                                //create help
+                                return false;
+                            }
                             languages = args[1].toLowerCase();
                             if (!abstractLanguageAPI.getAvailableLanguages().contains(languages)) {
                                 abstractLanguageAPI.createLanguage(languages);
                                 player.sendMessage(abstractLanguageAPI.getMessage("languageapi-create-success", player.getUniqueId(), true).replace("%LANG%", languages));
+                                return true;
                             } else {
-                                player.sendMessage(abstractLanguageAPI.getMessage("languageapi-languages-already-exists", player.getUniqueId(), true).replace("%LANG%", languages));
+                                player.sendMessage(abstractLanguageAPI.getMessage("languageapi-languages-already-exists", player.getUniqueId(), true)
+                                        .replace("%LANG%", languages));
+                                return false;
                             }
-                            break;
 
                         case "delete":
+                            if(!(args.length >= 2)) {
+                                return false;
+                            }
                             languages = args[1].toLowerCase();
                             if (abstractLanguageAPI.getAvailableLanguages().contains(languages) && !abstractLanguageAPI.getDefaultLanguage().equalsIgnoreCase(languages)) {
                                 abstractLanguageAPI.deleteLanguage(languages);
-                                player.sendMessage(abstractLanguageAPI.getMessage("languageapi-delete-success", player.getUniqueId(), true).replace("%LANG%", languages));
+                                player.sendMessage(abstractLanguageAPI.getMessage("languageapi-delete-success",
+                                        player.getUniqueId(),
+                                        true)
+                                        .replace("%LANG%", languages));
+                                return true;
                             } else if (languages.equalsIgnoreCase("*")) {
                                 for (String langs : abstractLanguageAPI.getAvailableLanguages()) {
                                     abstractLanguageAPI.deleteLanguage(langs);
                                 }
+                                player.sendMessage(abstractLanguageAPI.getMessage("languageapi-delete-success",
+                                        player.getUniqueId(),
+                                        true)
+                                        .replace("%LANG%", languages));
+                                return true;
                             } else {
                                 player.sendMessage(abstractLanguageAPI.getMessage("languageapi-languages-not-found", player.getUniqueId(), true)
                                         .replace("%LANG%", languages));
+                                return false;
 
                             }
-                            break;
                         case "copy":
                             if (args.length >= 3) {
                                 String langfrom = args[1].toLowerCase();
@@ -122,6 +145,7 @@ public class LanguageCommand implements TabExecutor {
                                     player.sendMessage(abstractLanguageAPI.getMessage("languageapi-copy-success", player.getUniqueId(), true)
                                             .replace("%OLDLANG%", langfrom)
                                             .replace("%NEWLANG%", langto));
+                                    return true;
                                 } else {
                                     languages = langfrom;
                                     if (abstractLanguageAPI.getAvailableLanguages().contains(langfrom)) {
@@ -129,23 +153,27 @@ public class LanguageCommand implements TabExecutor {
                                     }
                                     player.sendMessage(abstractLanguageAPI.getMessage("languageapi-languages-not-found", player.getUniqueId(), true)
                                             .replace("%LANG%", languages));
+                                    return false;
 
                                 }
-
-
                             } else {
                                 player.sendMessage(abstractLanguageAPI.getMessage("languageapi-copy-help", player.getUniqueId(), true));
+                                return false;
 
                             }
-                            break;
                         case "param": //languages show key
+                            if(!(args.length >= 2)) {
+                                return false;
+                            }
                             key = args[1].toLowerCase();
                             if (!abstractLanguageAPI.hasParameter(key) || abstractLanguageAPI.getParameter(key).equalsIgnoreCase("")) {
                                 player.sendMessage(abstractLanguageAPI.getMessage("languageapi-key-has-no-param", player.getUniqueId(), true).replace("%KEY%", key));
-                                break;
+                                return false;
+
                             }
                             player.sendMessage(abstractLanguageAPI.getMessage("languageapi-show-success", player.getUniqueId(), true)
                                     .replace("%PARAM%", abstractLanguageAPI.getParameter(key)).replace("%KEY%", key));
+                            return true;
                         case "translations":
                             languages = args[1].toLowerCase();
                             if (abstractLanguageAPI.getAvailableLanguages().contains(languages)) {
@@ -154,12 +182,14 @@ public class LanguageCommand implements TabExecutor {
                                     player.sendMessage(abstractLanguageAPI.getMessage("languageapi-translation-success", player.getUniqueId(), true)
                                             .replace("%KEY%", allKeys.get(i))
                                             .replace("%MSG%", abstractLanguageAPI.getAllTranslations(languages).get(i)));
+
                                 }
+                                return true;
                             } else {
                                 player.sendMessage(abstractLanguageAPI.getMessage("languageapi-languages-not-found", player.getUniqueId(), true)
                                         .replace("%LANG%", languages));
+                                return false;
                             }
-                            break;
                         case "remove": //languages remove languages key
                             if (args.length >= 3) {
                                 languages = args[1].toLowerCase();
@@ -170,6 +200,7 @@ public class LanguageCommand implements TabExecutor {
                                         player.sendMessage(abstractLanguageAPI.getMessage("languageapi-remove-key-in-languages", player.getUniqueId(), true)
                                                 .replace("%KEY%", key)
                                                 .replace("%LANG%", languages));
+                                        return true;
                                     } else if (key.endsWith("*")) {
                                         for (String keys : abstractLanguageAPI.getAllTranslationKeys(languages)) {
                                             if (keys.startsWith(key.replace("*", ""))) {
@@ -179,10 +210,12 @@ public class LanguageCommand implements TabExecutor {
                                         player.sendMessage(abstractLanguageAPI.getMessage("languageapi-remove-every-key-in-languages", player.getUniqueId(), true)
                                                 .replace("%LANG%", languages)
                                                 .replace("%STARTSWITH%", key.replace("*", "")));
+                                        return true;
                                     } else {
                                         player.sendMessage(abstractLanguageAPI.getMessage("languageapi-key-not-found", player.getUniqueId(), true)
                                                 .replace("%KEY%", key)
                                                 .replace("%LANG%", languages));
+                                        return false;
                                     }
                                 } else if (languages.equalsIgnoreCase("*")) {
                                     if (key.endsWith("*")) { //JEDE SPRACHE JEDER KEY
@@ -202,12 +235,27 @@ public class LanguageCommand implements TabExecutor {
                                         player.sendMessage(abstractLanguageAPI.getMessage("languageapi-remove-key-in-every-languages", player.getUniqueId(), true)
                                                 .replace("%KEY%", key));
                                     }
+                                    return true;
                                 }
                             }
                             break;
+                        case "reload":
+                            mySQL.closeConnection(); //KILL MYSQL
+                            mySQL.connect(); //START MYSQL
+                            Source.createSpigotMySQLConfig();
+                            I18N.createDefaultPluginMessages();
+                            Source.initSpigot();
+                            Source.defaultLanguage = null;
+                            player.sendMessage(abstractLanguageAPI.getMessage("languageapi-reload-success", player.getUniqueId(), true));
+                        default:
+                            abstractLanguageAPI.getMultipleMessages("languageapi-help", player.getUniqueId())
+                                    .forEach((s1 -> player.sendMessage(abstractLanguageAPI.getMessage(s1, player.getUniqueId(), true))));
+                            break;
                     }
                 } else {
-                    abstractLanguageAPI.getMultipleMessages("languageapi-help", player.getUniqueId()).forEach((s1 -> player.sendMessage(abstractLanguageAPI.getMessage(s1, player.getUniqueId(), true))));
+                    abstractLanguageAPI.getMultipleMessages("languageapi-help", player.getUniqueId())
+                            .forEach((s1 -> player.sendMessage(abstractLanguageAPI.getMessage(s1, player.getUniqueId(), true))));
+                    return false;
                 }
             }
         }
@@ -218,23 +266,23 @@ public class LanguageCommand implements TabExecutor {
     @Override
     public List<String> onTabComplete(@NotNull CommandSender commandSender, @NotNull Command command, @NotNull String s, String[] args) {
         if (args.length == 1) {
-            return this.getCompletes(args[0], tabComplete);
+            return this.getTabCompletes(args[0], tabComplete);
         } else if (args.length == 2) {
             if (!args[0].equalsIgnoreCase("param")) {
-                return this.getCompletes(args[1], abstractLanguageAPI.getAvailableLanguages());
+                return this.getTabCompletes(args[1], abstractLanguageAPI.getAvailableLanguages());
             }
-            return this.getCompletes(args[1], abstractLanguageAPI.getAllTranslationKeys(abstractLanguageAPI.getDefaultLanguage()));
+            return this.getTabCompletes(args[1], abstractLanguageAPI.getAllTranslationKeys(abstractLanguageAPI.getDefaultLanguage()));
         } else if (args.length == 3) {
             List<String> keyComplete = Arrays.asList("add", "remove", "update");
             if (keyComplete.contains(args[0].toLowerCase())) {
-                return this.getCompletes(args[2], abstractLanguageAPI.getAllTranslationKeys(args[1].toLowerCase()));
+                return this.getTabCompletes(args[2], abstractLanguageAPI.getAllTranslationKeys(args[1].toLowerCase()));
             }
         }
         return Collections.emptyList();
 
     }
 
-    private List<String> getCompletes(String playerInput, List<String> tabComplete) {
+    private List<String> getTabCompletes(String playerInput, List<String> tabComplete) {
         List<String> possibleCompletes = new ArrayList<>();
         StringUtil.copyPartialMatches(playerInput, tabComplete, possibleCompletes);
         Collections.sort(possibleCompletes);
